Move commonly used functions to util file
